requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
Pillow==9.5.0